fix: render shaders at display resolution instead of static 4x multiplier
Shaders now render at the actual window/canvas size rather than a fixed
4x multiple of the game's native resolution. The Compose canvas reports
its dimensions via gpuResize(), and the GPU renderer uses those to size
the offscreen target (maintaining game aspect ratio). This ensures shader
effects like LCD grid, scanlines, and CRT have proper pixel density at
any window size.

Key changes:
- Add desired_output_width/height to GPU renderer, set via gpuResize()
- render_to_bgra and HW render offscreen use window size for shaders
- Add output_size to push constants (std430 aligned) for all shaders
- JNI nativeGpuRenderToBgra returns packed width|height (jlong)
- MetalOffscreenSurface reports canvas size via onSizeChanged
- Update descriptor set when shader changes (sampler nearest↔linear)

+++ player/shared/src/desktopMain/kotlin/com/spela/player/libretro/LibretroJni.kt
@@ -66,7 +66,7 @@ class LibretroJni {
     external fun nativeGpuInit(surface: Any): Boolean
     external fun nativeGpuInitOffscreen(width: Int, height: Int): Boolean
     external fun nativeGpuRender()
-    external fun nativeGpuRenderToBgra(outData: ByteArray): Int
+    external fun nativeGpuRenderToBgra(outData: ByteArray): Long
     external fun nativeGpuSetShader(shaderId: Int)
     external fun nativeGpuResize(width: Int, height: Int)
     external fun nativeGpuDeinit()
